+++ vllm/model_executor/layers/attention.py
@@ -10,6 +10,12 @@ from xformers.ops.fmha.attn_bias import (BlockDiagonalCausalMask,
 from vllm import attention_ops
 from vllm import cache_ops
 from vllm import pos_encoding_ops
+from vllm.model_executor.layers.rotary_embedding import (
+    RotaryEmbedding,
+    LinearScalingRotaryEmbedding,
+    DynamicNTKScalingRotaryEmbedding
+)
+
 from vllm.model_executor.input_metadata import InputMetadata
 
 _SUPPORTED_HEAD_SIZES = [64, 80, 96, 112, 128, 256]
@@ -246,90 +252,6 @@ class PagedAttention(nn.Module):
         return output.view(-1, self.num_heads * self.head_size)
 
 
-class LlamaRotaryEmbedding(nn.Module):
-
-    def __init__(self,
-                 rotary_dim,
-                 max_position_embeddings=2048,
-                 base=10000) -> None:
-        super().__init__()
-        self.max_position_embeddings = max_position_embeddings
-        self.base = base
-        self.rotary_dim = rotary_dim
-
-        # Create the cos and sin cache.
-        inv_freq = 1.0 / (base**(torch.arange(0, rotary_dim, 2, dtype=torch.float, device='cuda') / rotary_dim))
-        self.register_buffer("inv_freq", inv_freq, persistent=False)
-        self._set_cos_sin_cache(seq_len=max_position_embeddings)
-
-    def _set_cache(self, t):
-        freqs = torch.einsum("i,j -> ij", t, self.inv_freq)
-        cos = freqs.cos()
-        sin = freqs.sin()
-        cache = torch.cat((cos, sin), dim=-1)
-        # FIXME(woosuk): This assumes that we configure the default dtype when
-        # initializing the model.
-        # TODO(woosuk): Make it more robust.
-        torch_dtype = torch.get_default_dtype()
-        cache = cache.to(torch_dtype)
-        self.register_buffer("cos_sin_cache", cache, persistent=False)
-
-    def _set_cos_sin_cache(self, seq_len):
-        self.max_seq_len_cached = seq_len
-        t = torch.arange(self.max_seq_len_cached, device='cuda')
-        self._set_cache(t)
-
-    # Unlink HF or TGI, we do not dynamically set the cos_sin_cache
-    # because vLLM needs to statically allocate memory in the beginning
-    # Therefore, we don't need to know the current seq_len
-    def forward(self):
-        return self.cos_sin_cache
-
-
-class LlamaLinearScalingRotaryEmbedding(LlamaRotaryEmbedding):
-
-    def __init__(self,
-                 rotary_dim,
-                 max_position_embeddings=2048,
-                 base=10000,
-                 scaling_factor=1.0):
-        self.scaling_factor = scaling_factor
-        super().__init__(rotary_dim, max_position_embeddings, base)
-
-    def _set_cos_sin_cache(self, seq_len):
-        self.max_seq_len_cached = seq_len * self.scaling_factor
-        t = torch.arange(self.max_seq_len_cached, dtype=torch.float, device='cuda') / self.scaling_factor
-        self._set_cache(t)
-
-
-class LlamaDynamicNTKScalingRotaryEmbedding(LlamaRotaryEmbedding):
-
-    def __init__(self,
-                 rotary_dim,
-                 max_position_embeddings=2048,
-                 base=10000,
-                 scaling_factor=1.0):
-        self.scaling_factor = scaling_factor
-        super().__init__(rotary_dim, max_position_embeddings, base)
-
-    def _set_cos_sin_cache(self, seq_len):
-        self.max_seq_len_cached = seq_len * self.scaling_factor
-        if seq_len > self.max_position_embeddings:
-            base = self.base * (
-                (self.scaling_factor * seq_len / self.max_position_embeddings)
-                - (self.scaling_factor - 1))**(self.rotary_dim /
-                                               (self.rotary_dim - 2))
-
-            inv_freq = 1.0 / \
-                (base **
-                 (torch.arange(
-                     0, self.rotary_dim, 2).float() / self.rotary_dim))
-            self.register_buffer("inv_freq", inv_freq, persistent=False)
-
-        t = torch.arange(self.max_seq_len_cached).float()
-        self._set_cache(t)
-
-
 class PagedAttentionWithRoPE(PagedAttention):
     """PagedAttention with GPT-NeoX style rotary embedding."""
 
@@ -349,13 +271,13 @@ class PagedAttentionWithRoPE(PagedAttention):
         self.is_neox_style = is_neox_style
         
         if rope_scaling is None:
-            self.rotary_emb = LlamaRotaryEmbedding(rotary_dim, max_position,
+            self.rotary_emb = RotaryEmbedding(rotary_dim, max_position,
                                                    base)
         elif rope_scaling["type"] == "linear":
-            self.rotary_emb = LlamaLinearScalingRotaryEmbedding(
+            self.rotary_emb = LinearScalingRotaryEmbedding(
                 rotary_dim, max_position, base, rope_scaling["factor"])
         elif rope_scaling["type"] == "dynamic":
-            self.rotary_emb = LlamaDynamicNTKScalingRotaryEmbedding(
+            self.rotary_emb = DynamicNTKScalingRotaryEmbedding(
                 rotary_dim, max_position, base, rope_scaling["factor"])
         else:
             raise ValueError(rope_scaling)
